bot-reserva/requirements.txt
botbuilder-core>=4.14.0
botbuilder-ai>=4.14.0
botbuilder-dialogs>=4.14.0
botbuilder-schema>=4.14.0
botbuilder-applicationinsights>=4.14.0
botbuilder-integration-aiohttp>=4.14.0
requests>=2.25.1
aiohttp>=3.7.4
azure-ai-language-conversations